Fixed promote.sh script

+++ promote.sh
@@ -5,10 +5,15 @@ if [ "$#" -ne 1 ]; then
     exit 1
 fi
 
+if [ ! -d "p2updatesite" ]; then
+    echo "Missing directory: p2updatesite"
+    exit 2
+fi
+
 BUILD=$1
 
-SRC=/opt/public/jobs/cbi_henshin_$BUILD/workspace/p2updatesite/target/repository
-TRG=/home/data/httpd/download.eclipse.org/modeling/emft/henshin/updates/$BUILD
+SRC="p2updatesite/target/repository"
+TRG="/home/data/httpd/download.eclipse.org/modeling/emft/henshin/updates/$BUILD"
 
 rm -R $TRG/* 2> /dev/null
 cp -R $SRC/* $TRG
